+++ projector/expression.py
@@ -1,4 +1,4 @@
-import projector_error as error
+import projector.error as error
