+++ projector/token.py
@@ -1,4 +1,4 @@
-import projector_constants as constants
+import projector.constants as constants
