+++ tests/tests.py
@@ -23,9 +23,6 @@ class EvaluationTests(unittest.TestCase):
         self.assertEqual(evaluate("5*(2+1)"), 15)
         self.assertEqual(evaluate("5+(2*(4/2))-1"), 8)
 
-    def test_variables(self):
-        self.assertEqual(evaluate("x = 5"))
-
